fix: 회원 탈퇴 기능 auth-> user 로 도메인 이동
