+++ src/main/java/devping/nnplanner/domain/auth/service/AuthService.java
@@ -6,20 +6,11 @@ import devping.nnplanner.domain.auth.entity.User;
 import devping.nnplanner.domain.auth.entity.User.LoginType;
 import devping.nnplanner.domain.auth.repository.EmailRepository;
 import devping.nnplanner.domain.auth.repository.UserRepository;
-import devping.nnplanner.domain.monthmenu.entity.MonthMenu;
-import devping.nnplanner.domain.monthmenu.entity.MonthMenuHospital;
-import devping.nnplanner.domain.monthmenu.entity.MonthMenuSchool;
-import devping.nnplanner.domain.monthmenu.repository.MonthMenuHospitalRepository;
-import devping.nnplanner.domain.monthmenu.repository.MonthMenuRepository;
-import devping.nnplanner.domain.monthmenu.repository.MonthMenuSchoolRepository;
-import devping.nnplanner.domain.survey.entity.Survey;
-import devping.nnplanner.domain.survey.repository.SurveyRepository;
 import devping.nnplanner.global.exception.CustomException;
 import devping.nnplanner.global.exception.ErrorCode;
 import devping.nnplanner.global.jwt.token.JwtUtil;
 import devping.nnplanner.global.jwt.user.UserDetailsImpl;
 import jakarta.servlet.http.HttpServletRequest;
-import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.crypto.password.PasswordEncoder;
@@ -35,10 +26,6 @@ public class AuthService {
     private final UserRepository userRepository;
     private final EmailRepository emailRepository;
     private final JwtUtil jwtUtil;
-    private final MonthMenuRepository monthMenuRepository;
-    private final MonthMenuHospitalRepository monthMenuHospitalRepository;
-    private final MonthMenuSchoolRepository monthMenuSchoolRepository;
-    private final SurveyRepository surveyRepository;
 
     @Transactional
     public void signUp(AuthSignRequestDTO authSignRequestDTO) {
@@ -90,36 +77,4 @@ public class AuthService {
         jwtUtil.deleteRefreshToken(userId, email);
         jwtUtil.logoutAccessToken(httpRequest);
     }
-
-    public void signOut(UserDetailsImpl userDetails) {
-
-        User user = userRepository.findById(userDetails.getUser().getUserId())
-                                  .orElseThrow(() -> new CustomException(ErrorCode.NOT_FOUND));
-
-        List<MonthMenu> monthMenus =
-            monthMenuRepository.findAllByUser_UserId(userDetails.getUser().getUserId());
-
-        for (MonthMenu monthMenu : monthMenus) {
-
-            List<MonthMenuHospital> monthMenuHospitals =
-                monthMenuHospitalRepository.findAllByMonthMenu_MonthMenuId(
-                    monthMenu.getMonthMenuId());
-            monthMenuHospitalRepository.deleteAll(monthMenuHospitals);
-
-            List<MonthMenuSchool> monthMenuSchools =
-                monthMenuSchoolRepository.findAllByMonthMenu_MonthMenuId(
-                    monthMenu.getMonthMenuId());
-            monthMenuSchoolRepository.deleteAll(monthMenuSchools);
-
-            List<Survey> surveys =
-                surveyRepository.findAllByMonthMenu_MonthMenuId(monthMenu.getMonthMenuId());
-            surveyRepository.deleteAll(surveys);
-
-            monthMenuRepository.delete(monthMenu);
-        }
-        
-        userRepository.delete(user);
-
-    }
-
 }
